correct folder structure and file names
