+++ client/src/Router.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import {BrowserRouter, Route, Switch} from 'react-router-dom'
-import Home2 from './components/Version 2/Home2'
+import Home from './components/Home'
 
 export default class Router extends Component {
   render() {
@@ -8,8 +8,7 @@ export default class Router extends Component {
       <BrowserRouter>
         <Switch>
           {/* this matches everything unless u add 'exact' prop */}
-          {/* <Route path='/old' component={Home} /> */}
-          <Route path='/' component={Home2} />
+          <Route path='/' component={Home} />
         </Switch>
       </BrowserRouter>
     )
