+++ client/src/components/NewNote.js
@@ -1,136 +1,117 @@
-import React from 'react'
-import {Mutation} from 'react-apollo'
-import {addNoteMutation, getAllNotesQuery} from '../queries/index'
+import React, {useState} from 'react'
 import styled from 'styled-components'
+import {addNoteRequest} from '../requests'
 
-export default class NewNote extends React.Component{
-    state = {title: "", text:"", disabled:true}
-    
-    // let user type into the box
-    handleChange = (event)=> {
-        let titleOrText = event.target.getAttribute('name')
+function NewNote (props) {
+	const [note, setNote] = useState({title: '', text: ''})
+	const [isDisabled, setDisabled] = useState(true)
+	const {notes, setNotes, showTitle, setShowTitle, darkTheme} = props
 
-        this.setState({[titleOrText]: event.target.innerHTML})
+	const handleAddNoteArgs = {
+		note, setNote, setShowTitle, setDisabled, notes, setNotes
+	}
 
-        const currentTitle = document.getElementById('title').innerHTML
-        const currentText = document.getElementById('text').innerHTML
-
-        // disable add note button if no text AND no title
-        if (currentTitle.length === 0 && currentText.length === 0) this.setState({disabled: true})
-        else this.setState({disabled: false})
-    }
-    
-    // turn on title input box
-    handleClick = (event) => {
-        // prevent confliction with handleClick on the parent component
-        event.stopPropagation()
-        if(!this.props.displayTitleInputBox)
-            this.props.turnOnTitleBox()
-    }
-
-    // send graphql mutation
-    handleAddNote = (mutate, event)=> {
-        event.preventDefault()
-
-        const title = document.getElementById('title').innerHTML
-        let text = document.getElementById('text').innerHTML
-        
-        if (text.length > 0) {
-            text = text.replace(/<div>/gi, '\n');
-            text = text.replace(/<\/div>/gi, '');
-            text = text.replace(/<br>/gi, '');
-        }
-
-        // clear form and hide title box
-        this.setState({title: '', text: ''}, ()=> this.props.turnOffTitleBox())
-        
-        // state and div innerHTML are no longer linked so need to manually clear each
-        // linking them makes typing the text go the wrong way
-        document.getElementById('title').innerHTML = ''
-        document.getElementById('text').innerHTML = ''
-        
-        // disable button
-        this.setState({disabled: true})
-
-        // make request
-        mutate(
-            {variables: 
-                { input: 
-                    {
-                        "title": title,
-                        "text": text
-                    } 
-                }
-            }
-        )
-        // close title box
-    }
-
-    // update apollo cache when user adds note to reflect changes immediately on UI
-    update = (cache, result)=> {
-        let {allNotes} = cache.readQuery({query: getAllNotesQuery })
-        cache.writeQuery({
-            query: getAllNotesQuery,
-            data: {allNotes: [result.data.addNote, ...allNotes]}
-        })
-    }
-
-    // to prevent line breaks in title box
-    preventLineBreak = (e)=> {
-
-        // prevent 'enter' key from creating line break in title box
-        if (e.keyCode === 13) {
-            e.preventDefault()
-            // instead have it to go to text box to simulate a tab press
-            document.querySelector('#text').focus()
-        }
-    }
-
-    createTitleBox = () => {
-        return this.props.displayTitleInputBox ? 
-        <AddNoteInput id='title' name='title' suppressContentEditableWarning={true}
-        contentEditable data-placeholder='Add a title...' 
-        onInput={this.handleChange}
-        onKeyDown={this.preventLineBreak}>
-        {/* should be empty otherwise it reverses the text direction */}
-        </AddNoteInput> : null
-    }
-
-    createTextBox = () => {
-        return (
-            <AddNoteInput id='text' name='text' contentEditable
-                suppressContentEditableWarning={true} data-placeholder='Add a note...'
-                onInput={this.handleChange}>{/* should be empty otherwise it reverses text direction */}
-            </AddNoteInput>
-        )
-    }
-
-    render() {
-        return (
-            <Mutation mutation={addNoteMutation}
-                update={this.update}>
-            {
-                (mutate, result) => {
-                    return (
-                        <AddNoteContainer onClick={this.handleClick} autoComplete='off' darkTheme={this.props.darkTheme}>
-                            {this.createTitleBox()}
-                            {this.createTextBox()}
-                            {
-                                this.props.displayTitleInputBox ?
-                                <AddNoteButton type='submit' disabled={this.state.disabled}
-                                onClick={this.handleAddNote.bind(this, mutate)}
-                                darkTheme={this.props.darkTheme}>Add</AddNoteButton> : null
-                            }
-                        </AddNoteContainer>
-                    )
-                }
-            }
-            </Mutation>
-        )
-    }
+	return ( 
+		<AddNoteContainer id='addNoteContainer'
+		onClick={e => revealTitle(setShowTitle, e)}
+		darkTheme={darkTheme}
+		autoComplete='off'>
+			<TitleInput setDisabled={setDisabled}
+			showTitle={showTitle} note={note}
+			setNote={setNote}/>
+			<TextInput setDisabled={setDisabled} 
+			setNote={setNote} note={note} />
+			{
+				showTitle ? 
+				<AddNoteButton id='addnote'
+				type='submit' disabled={isDisabled}
+				onClick={() => handleAddNote(handleAddNoteArgs)} 
+				darkTheme={darkTheme}>Add
+				</AddNoteButton> : null
+			}
+		</AddNoteContainer>
+	)
 }
 
-export const AddNoteContainer = styled.div`
+function revealTitle (setShowTitle, e) {
+	const id = e.nativeEvent.target.id
+	// ignore add note button to avoid collision
+	if(id === 'addnote') return
+	setShowTitle(true)
+}
+
+async function handleAddNote ({note, setNote, setShowTitle, setDisabled, notes, setNotes}) {
+	let {title, text} = note
+	text = sanitize(text)
+
+	// state and div innerHTML are no longer linked so need to manually clear each, linking them makes typing the text go the wrong way
+	document.getElementById('title').innerHTML = ''
+	document.getElementById('text').innerHTML = ''
+
+	// clear state, hide title and disable button
+	setNote({title: '', text: ''})		
+	setDisabled(true)
+	setShowTitle(false)
+
+	const {data: {newNote}} = await addNoteRequest({title, text})
+
+	// update ui
+	const updatedNotes = [newNote, ...notes] // must create new object(therefore new reference) or else react won't know state has changed and will bail out without rendering children or firing effects, this is because for OBJECTS, react compares references instead of values
+	setNotes(updatedNotes)
+}
+
+
+function TitleInput({note, setNote, setDisabled, showTitle}) {
+	return showTitle ? 
+	<AddNoteInput id='title' name='title' suppressContentEditableWarning={true}
+    contentEditable data-placeholder='Add a title...' 
+    onInput={updateState.bind(this, note, setNote, setDisabled)}
+    onKeyDown={preventLineBreak}/> : null
+}
+
+function TextInput({note, setNote, setDisabled}) {
+	return (
+		<AddNoteInput id='text' name='text' contentEditable
+			suppressContentEditableWarning={true} data-placeholder='Add a note...'
+			onInput={updateState.bind(this, note, setNote, setDisabled)}/>
+	)
+}
+
+
+
+function sanitize(text) {
+	if (text.length > 0) {
+		text = text.replace(/<div>/gi, '\n');
+		text = text.replace(/<\/div>/gi, '');
+		text = text.replace(/<br>/gi, '\n');
+		text = text.replace(/&nbsp/gi, ' ');
+	}
+	return text
+}
+
+
+function updateState (note, setNote, setDisabled, event) {
+	let titleOrText = event.target.getAttribute('name')
+	const val = event.target.innerHTML
+	note = {...note,  [titleOrText]: val}
+	setNote(note)
+
+	// disable add note button if no text AND no title
+	if (note.title.length === 0 && note.text.length === 0) 
+		setDisabled(true)
+	else setDisabled(false)
+}
+
+function preventLineBreak (e) {
+	// prevent 'enter' key from creating line break in title box
+	if (e.keyCode === 13) {
+		e.preventDefault()
+		// instead go to TextInputBox
+		document.querySelector('#text').focus()
+	}
+}
+
+const AddNoteContainer = styled.div`
     background: ${props => props.darkTheme ? `rgb(40, 40, 40)` : `white`};
     border-radius: 5px;
     box-shadow: 0px 1px 3px rgba(0, 0, 0, 0.05) inset, 0px 0px 8px ${props => props.darkTheme ? `white` : `black`};
@@ -150,7 +131,7 @@ export const AddNoteContainer = styled.div`
 `
     // border: solid 1px;
 // TODO: change this to span to prevent submission of empty notes that have been cleared after typing some characters
-export const AddNoteInput = styled.span`
+const AddNoteInput = styled.span`
     border:none;
     width:96%;
     border-radius: 10px;
@@ -166,7 +147,7 @@ export const AddNoteInput = styled.span`
     :focus {outline: 0;}
 `
 
-export const AddNoteButton = styled.button`
+const AddNoteButton = styled.button`
     cursor: ${props => props.disabled ? `normal` : `pointer`};
     background: inherit;
     border-radius: 10px;
@@ -186,3 +167,11 @@ export const AddNoteButton = styled.button`
         }`
     }
 `
+
+function areEqual (prevProps, nextProps) {
+	if (prevProps.showTitle !== nextProps.showTitle) return false
+	if (prevProps.darkTheme !== nextProps.darkTheme) return false
+	return true
+}
+
+export default React.memo(NewNote, areEqual)
