+++ client/src/components/Note.js
@@ -1,31 +1,14 @@
 import React from 'react'
 import styled from 'styled-components'
 
-export default class Note extends React.Component {
-
-    showModal = (note, event) => {
-        let autoFocusLocation
-        if (event.target.id === 'noteTitle')
-            autoFocusLocation = 'title'
-        else autoFocusLocation = 'text'
-
-        this.props.showModal(note, autoFocusLocation)
-        
-        // hide original note box while modal is here
-        let noteBox = document.getElementById(note.id)
-        noteBox.style.visibility = 'hidden'
-    }
-
-    render() {
-        const note = this.props.note
-        note.text = note.text.replace(/\n/g, '<br/>');
-        return (
-            <NoteBox id={note.id} onClick={this.showModal.bind(this, note)} darkTheme={this.props.darkTheme}>
-                <NoteTitle id='noteTitle'  dangerouslySetInnerHTML={{__html: note.title}}></NoteTitle>
-                <NoteText id='noteText' dangerouslySetInnerHTML={{__html: note.text}}></NoteText>
-            </NoteBox>
-        )
-    }
+export default function Note({note: {id, title, text}, openNoteInModal, darkTheme}) {
+	
+	return (
+		<NoteBox id={id} darkTheme={darkTheme} onClick={e => openNoteInModal(e, {id, title, text})}>
+			<NoteTitle id='noteTitle'>{title}</NoteTitle>
+			<NoteText id='noteText'>{text}</NoteText>
+		</NoteBox>
+	)
 }
 
 const NoteBox = styled.div`
@@ -33,15 +16,19 @@ const NoteBox = styled.div`
     border-radius: 5px;
     margin: 3% 1% 3% 1%;
     padding: 3%;
-    display: flex;
+	display: flex;
+	// word-wrap: anywhere;
+	word-break: break-word;
     flex-direction: column;
     justify-content: flex-start;
-    word-break: break-all;
     @media only screen and (min-width: 1080px) {
-        margin: 5%;
-    }
+		margin: 3%;
+	}
+	border: solid 1px ${props => props.darkTheme ? `#666666` : `#111111`};
+	:hover {
+		box-shadow: 0px 1px 4px ${props => props.darkTheme ? `white` : `black`}
+	}
 `
-
 const NoteTitle = styled.span`
     font-size: 20px;
     font-weight: bold;
